Added run and MsgBox support with ~ and #
You can now dynamically run functions with a flag in your hotkey action by including the name of the function and a flag in line 9 of the config.ini file comma separated. For example ~:run,#:MsgBox means that the run function will be called with whatever else is in the hotstring  as an argument. This means hotkeys can now get really fancy.